[YIYO-9]: Create Entity and ValueObject
In this commit, we first implement the `Entity` (thanks [Microsoft](https://learn.microsoft.com/en-us/dotnet/architecture/microservices/microservice-ddd-cqrs-patterns/seedwork-domain-model-base-classes-interfaces))
and `ValueObject` classes.
We start with the `Entity`. It has an `Id` with a protected set, so that anything extending `Entity` can define its
id, but I might add a little trick later, where you can only set it if it's 0. For now, it's
an idea to remember.
There is an issue though...
Next is `ValueObject`, and we can have a similar one to the one microsoft suggest, just with a little else.