[YIYO-9]: Create entities for Domain.Reflection
In this commit we create the Entities folder in the Models folder and populate it with the rest of
the entities from the PlantUML diagram.
I've added a `Title` property to `ReflectionEvent` and a `Title` and `Description` to the `Feeling`.
I also decided to have them mutable for now, so the public properties are `{ get; set; }`