[YIYO-16]: Create unit tests for GetFeelingQuery
In this commit, we prep the unit testing of `GetFeelingQuery`'s handler.
We prepped the query itself with an Id, and an internal handler having an `IApplicationDbContext`
with the Handle not implemented.
We then create the test suite with a mock of `IApplicationDbContext`, and the handler.
We also need to create a test case source that will supply the mock rows of feelings to the
mock at setup and then sends the correct expected stuff for the test cases.
We will allow several options in the query, but I am not sure whether this is the correct approach. I will ask for
feedback from people.

We also add FluentAssertions to the unit test project.
We make the test case source with `FeelingEntity`, and also everything related to the query.