[YIYO-16]: Create more unit tests covering Feeling's title
In this commit, we create two more tests to cover an existing and non existant title
in the Feelings DbSet.
We additionally refactor the mocking of the DbSet to a method, which we might later pull up
to a test base class to allow it to be reused for the rest of the entities.