[YIYO-16]: Create unit tests for GetMottoQuery
In this commit, we create the unit tests for the motto queries.
On the way, we create a `TestFixtures` class where we put the `DbSetMock` method
and we generalize it so it is usable for all entities.
We also changed `IDbEntity` to public to be used as a constraing for `DbSetMock`
type.