[YIYO-18] fix: Fix the DbSet mock to allow addition
In this commit, we add a callback to the `DbSetMock` method to add to the original list when trying
to add to the `DbSet` mock. In order for this to work though, we need the parameter to be not read only
and therefore we put an `IList`. We then fix the test cases source to use `IList`s as well.