[YIYO-18] fix(testing): Fix AddFeelingCommand tests regarding invalid feelings
In this commit, we fix the validation of Feeling to throw FeelingException for invalid
Title and Exception, adjust the TestCaseSource and the assertion of the test to throw
the exception. In addition, we add a simple test for null feeling and for feelings
with none zero ids.
Technically, that makes the test for existing feelings currently redundant, since it
is currently implements such that non zero ids return early, but it should
be there should a change that breaks this is attempted.