[YIYO-18] feat: Add validation to all domain entities
Before we continue with the rest of the addition commands,
we will establish validation on all domain entities.
However, I've decided to retreat from a separate exception class
per entity, and will instead have one with some type for which
entity type it is which will be incorporated into the message.

As I was validating the `WorldEvent` I came accross an issue.
what do we do with `Source`, which inherits from `ValueObject`?
Should it have its own `Validate`? Or should `WorldEvent` do the validation?
And that added more questions:
Does a source only gets added if a world event is added? Should the `SourceEntity`
only have a string `Url` which is only ever converted to `Uri` in the Domain layer?
I ended up making the SourceEntity have a string and then the validation is part of
the world event's duty.
The logic behind it is that the sources are secondary to the world events, and not what
the user cares about adding, from a UX perspective. And therefore, the business logic
should reflect that about the domain.
The Source has no business logic of its own, and only its owner, the WorldEvent
should validate it when it is validated.