[YIYO-18] fix(testing): Replace DbSet mocks using MockQueryable.Moq package
In this commit, we replace the mocking of DbSets using a package called
`MockQueryable.Moq` in order to allow mocking of async operations (and in general
more stable mocking). This of course means I only need to setup add & add async for the `DbSet`
mock, which I will reduce the `TestFixtures.DbSet` to.
We also fix an issue with undesired reference equality in the feeling command test AND
accidentally having side effects by changing the test case sources directly.
We also fix some incorrect test names