[YIYO-18] fix(mappers): Add a default MapMany to the domain to db mapper interface
I did another big oh no - I did not map back the lists from domain to db. I will change the properties to be
`IList` rather than `ICollection` to guarantee order so that the different lists will be **guaranteed**
to be ordered the same.
First, Let's add a fix for the interface, then a fix for the existing mappers before adding the missing ones.

Additional fixes:
- Change from `ICollection` to `IList` in db entities
- I had to add another line to the callback in the AddWorldEvent tests since it was saving a source with id 0.