[YIYO-18] refactor: Source commands in WorldEvent command replace with IMediator
In this commit, we replace the explicit handlers for the sources in the
AddWorldEventRequestHandler with an IMediator that send the correct commands.
I also move the bool for whether it is a child command to the command itself
since I don't control the handler construction anymore (and besides it makes
more sense there anyways).
I also update the tests accordingly