[34] fix: Change dev connection string and create missing tables
I had to remove two things from the connection string of appsettings.Development.json:
"Trusted_Connection=True;MultipleActiveResultSets=true".
We installed the PostgreSQL locally, and then changed the host to localhost:5432.
I also had to add a section in the Persistence layer DI that ensures the database
is created and creates the table if necessary.
I will add a boolean of `HasTables` before calling the `CreateTables`.