[14]: Create extension methods for ControllerBase controller route
As part of the Created response I want to return the URI as REST would
want me to. For that I need the URL of the controller. FULL URL.
As I explore my options, I realize I will need it in all controllers.
I briefly flirt with the idea of having a superclass, but I then realize
that it is smarter to just to an extension method to ControllerBase, since
all my controllers already inherit from it, and having a class in between
just for a couple of utility methods is not necessary.
So I create extension methods to retrieve the controller route.