[18]: Add FeelingCommandValidator
In this commit we add `FeelingCommandValidator` which extends `CommandValidator<FeelingEntity>`.
In order to facilitate getting all the db PersonalEvent, I add a new
query `GetPersonalEventlistQuery` which accepts a list of ids and retrieves
the corresponding personal events. Or for a null list, retrieves all
of them.
In `GetPersonalEventListQuery` I decide to have a set for the ids.
The reason is that in the actual handler I will check for a predicate with
Ids containing the id. That would be O(N) for a list and O(1) for a set,
making it worth it to spend the O(N) of making a set when calling it.
We might need a query validator now, to check that all ids are non 0.
In order for `SequenceEquals` to work, we need to override Equals for the
`PersonalEventEntity`.